services

+++ src/business_logic/services/device_auth.py
@@ -2,7 +2,7 @@ import logging
 import random
 from string import ascii_uppercase
 import secrets
-
+from typing import Any, Optional
 from src.data_access.postgresql.repositories import (
     ClientRepository,
     DeviceRepository,
@@ -19,47 +19,60 @@ class DeviceService:
             client_repo: ClientRepository,
             device_repo: DeviceRepository,
     ) -> None:
-        self._request_model = None
+        self._request_model: Optional[RequestModel] = None
         self.client_repo = client_repo
         self.device_repo = device_repo
 
-    async def get_response(self) -> dict:
-        device_code = secrets.token_urlsafe(32)
-        user_code = "".join(random.sample(ascii_uppercase, k=8))
-        verification_uri = "http://127.0.0.1:8000/device/auth"
-        verification_uri_complete = f"http://127.0.0.1:8000/device/outh?user_code={user_code}"
-        if await self._validate_client(client_id=self.request_model.client_id):
-            device_data = {
-                "device_code": device_code,
-                "user_code": user_code,
-                "verification_uri": verification_uri,
-                "verification_uri_complete": verification_uri_complete,
-                "expires_in": 600,
-                "interval": 5
-            }
-            await self.device_repo.create(client_id=self.request_model.client_id, **device_data)
+    async def get_response(self) -> Optional[dict[str, Any]]:
+        if self.request_model is not None:
+            device_code = secrets.token_urlsafe(32)
+            user_code = "".join(random.sample(ascii_uppercase, k=8))
+            verification_uri = "http://127.0.0.1:8000/device/auth"
+            verification_uri_complete = f"http://127.0.0.1:8000/device/outh?user_code={user_code}"
+            if await self._validate_client(client_id=self.request_model.client_id):
+                device_data:dict[str, Any] = {
+                    "device_code": device_code,
+                    "user_code": user_code,
+                    "verification_uri": verification_uri,
+                    "verification_uri_complete": verification_uri_complete,
+                    "expires_in": 600,
+                    "interval": 5
+                }
+                await self.device_repo.create(client_id=self.request_model.client_id, **device_data)
 
-            return device_data
-
+                return device_data
+        return None
+    
     async def get_redirect_uri(self) -> str:
-        uri_start = "http://127.0.0.1:8000/authorize/?"
-        redirect_uri = "https://www.google.com/"
-        device = await self.device_repo.get_device_by_user_code(user_code=self.request_model.user_code)
-        final_uri = uri_start + f"client_id={device.client.client_id}" \
-                                f"&response_type=urn:ietf:params:oauth:grant-type:device_code" \
-                                f"&redirect_uri={redirect_uri}&scope=user_code={self.request_model.user_code}"
+        if self.request_model is not None:
+            uri_start = "http://127.0.0.1:8000/authorize/?"
+            redirect_uri = "https://www.google.com/"
+            if self.request_model.user_code is None:
+                raise ValueError
+            device = await self.device_repo.get_device_by_user_code(user_code=self.request_model.user_code)
+            final_uri = uri_start + f"client_id={device.client.client_id}" \
+                                    f"&response_type=urn:ietf:params:oauth:grant-type:device_code" \
+                                    f"&redirect_uri={redirect_uri}&scope=user_code={self.request_model.user_code}"
 
-        return final_uri
+            return final_uri
+        else:
+            raise ValueError
 
     async def clean_device_data(self) -> str:
+        if self.request_model is None:
+                raise ValueError
         if await self._validate_client(client_id=self.request_model.client_id):
-            scope_data = await self._parse_scope_data(scope=self.request_model.scope)
+            scope_data={}
+            if self.request_model.scope is None:
+                scope_data = {'scope':'no_scope'}
+            else:
+                scope_data = await self._parse_scope_data(scope=self.request_model.scope)
             user_code = scope_data["user_code"]
             if await self._validate_user_code(user_code=user_code):
                 await self.device_repo.delete_by_user_code(user_code=user_code)
         return "http://127.0.0.1:8000/device/auth/cancel"
 
-    async def _parse_scope_data(self, scope: str) -> dict:
+    async def _parse_scope_data(self, scope: str) -> dict[str, str]:
         """ """
         return {
             item.split("=")[0]: item.split("=")[1]
@@ -80,7 +93,7 @@ class DeviceService:
         return client
 
     @property
-    def request_model(self) -> None:
+    def request_model(self) -> Optional[RequestModel]:
         return self._request_model
 
     @request_model.setter
